fix issue of showing PenaltyAmount

+++ LibraryMS/Application Service/DTOs/UserDto.cs
@@ -10,5 +10,5 @@ public class UserDto
     public string Password { get; set; }
     public UserRoleEnum Roll { get; set; }
     public bool Status { get; set; }
-    public int PenaltyAmount { get; set; }
+    public decimal PenaltyAmount { get; set; }
 }
